feat(AddArticle): renames ArticlePath to ArticleUrl
