+++ Shared/articles/domain/ArticleUrl.swift
@@ -1,0 +1,32 @@
+//
+//  ArticlePath.swift
+//  dev-articles
+//
+//  Created by Mathias Remshardt on 31.07.21.
+//
+
+import Foundation
+
+struct ArticleUrl {
+  private static let pathRegexString = #"^(http(s)?:\/\/)?dev.to\/(?<path>[a-zA-z]*\/[a-zA-Z\-1]*)$"#
+
+  let path: String
+
+  init?(url: String) {
+    if let path = ArticleUrl.matchPath(url: url) {
+      self.path = path
+    } else {
+      return nil
+    }
+  }
+
+  private static func matchPath(url: String) -> String? {
+    guard let regex = try? NSRegularExpression(pattern: pathRegexString),
+          let match = regex.firstMatch(in: url, range: NSRange(location: 0, length: url.count)),
+          let pathIndex = Range(match.range(withName: "path"), in: url)
+    else {
+      return nil
+    }
+    return String(url[pathIndex])
+  }
+}
